Add files via upload
Step 1: Add the ShiftSupervisor Class
In a particular factory, a shift supervisor is a salaried employee who supervises a shift. In addition to a salary, the supervisor earns a yearly bonus when production goals are met.
Design a class named ShiftSupervisor that extends the Employee class. This class should include:

A field for the annual salary
A field for the annual production bonus earned
Write one or more constructors and include the appropriate accessor and mutator methods for each field.
Step 2: Add the TeamLeader Class
A team leader is an hourly paid production worker who leads a small team. In addition to hourly pay, team leaders earn a fixed monthly bonus and must attend a certain number of training hours each year.